Refactor navigation commands and tests; consolidate navigation logic and enhance logout functionality
