+++ cypress/e2e/auth/login.cy.ts
@@ -54,18 +54,4 @@ describe('Login Page', () => {
     cy.url().should('include', '/auth/login');
   });
 
-  /**
-   * @description Verifica el flujo completo de inicio y cierre de sesión
-   * @test Debe permitir al usuario iniciar sesión correctamente
-   * @test Debe mostrar el menú de usuario después del login
-   * @test Debe permitir cerrar sesión mediante el botón correspondiente
-   * @test Debe redireccionar a la página de login después del logout
-   */
-  it('should redirect to login page after logout', () => {
-    cy.login('rh');
-    cy.url().should('not.include', '/auth/login');
-
-    cy.logout();
-    cy.url().should('include', '/auth/login');
-  });
 });
