+++ cypress/support/navigation/navigation-to-user.ts
